requirements.txt
absl-py==0.7.0
alembic==1.0.10
astor==0.7.1
astroid==2.2.4
autopep8==1.4.3
Click==7.0
cycler==0.10.0
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
isort==4.3.12
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Mako==1.0.10
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
numpy==1.16.2
pandas==0.24.1
pbr==5.1.3
protobuf==3.7.0
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pylint==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
pytz==2018.9
PyYAML==5.1
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.1
Werkzeug==0.14.1
wrapt==1.11.1
requests==2.22.0
scikit-learn==0.20.2
